refactor: move operator read model tests
